+++ code/__DEFINES/preferences.dm
@@ -105,15 +105,15 @@
 
 // Values for /datum/preferences/current_tab
 /// Open the character preference window
-#define PREFERENCE_TAB_CHARACTER_PREFERENCES 0
+#define PREFERENCE_WINDOW_CHARACTERS 0
 
 /// Open the game preferences window
-#define PREFERENCE_TAB_GAME_PREFERENCES 1
+#define PREFERENCE_WINDOW_GAME_PREFERENCES 1
 
-#define PREFERENCE_TAB_GAME_PREFERENCES_CHARACTER -1
-#define PREFERENCE_TAB_GAME_PREFERENCES_SETTINGS 0
-#define PREFERENCE_TAB_GAME_PREFERENCES_KEY_BINDINGS 1
-#define PREFERENCE_TAB_GAME_PREFERENCES_VOLUME 2
+#define PREFERENCE_PAGE_CHARACTERS -1
+#define PREFERENCE_PAGE_SETTINGS 0
+#define PREFERENCE_PAGE_KEY_BINDINGS 1
+#define PREFERENCE_PAGE_PREFERENCES_VOLUME 2
 
 /// These will be shown in the character sidebar, but at the bottom.
 #define PREFERENCE_CATEGORY_FEATURES "features"
