+++ code/modules/client/preferences/_preference.dm
@@ -350,7 +350,7 @@ GLOBAL_LIST_INIT(preference_entries_by_key, init_preference_entries_by_key())
 /// Returns whether or not, given the PREFERENCE_TAB_*, this preference should
 /// appear.
 /datum/preference/proc/should_show_on_page(preference_tab)
-	var/is_on_character_page = preference_tab == PREFERENCE_TAB_CHARACTER_PREFERENCES
+	var/is_on_character_page = preference_tab == PREFERENCE_WINDOW_CHARACTERS
 	var/is_character_preference = savefile_identifier == PREFERENCE_CHARACTER
 	return is_on_character_page == is_character_preference
