+++ code/modules/client/preferences/jobless_role.dm
@@ -9,4 +9,4 @@
 	return list(BEOVERFLOW, BERANDOMJOB, RETURNTOLOBBY)
 
 /datum/preference/choiced/jobless_role/should_show_on_page(preference_tab)
-	return preference_tab == PREFERENCE_TAB_CHARACTER_PREFERENCES
+	return preference_tab == PREFERENCE_WINDOW_CHARACTERS
