+++ code/modules/client/preferences/middleware/antags.dm
@@ -4,7 +4,7 @@
 	)
 
 /datum/preference_middleware/antags/get_ui_static_data(mob/user)
-	if (preferences.current_window != PREFERENCE_TAB_CHARACTER_PREFERENCES)
+	if (preferences.current_window != PREFERENCE_WINDOW_CHARACTERS)
 		return list()
 
 	var/list/data = list()
