+++ code/modules/client/preferences/middleware/legacy_toggles.dm
@@ -40,7 +40,7 @@
 	)
 
 /datum/preference_middleware/legacy_toggles/get_character_preferences(mob/user)
-	if (preferences.current_window != PREFERENCE_TAB_GAME_PREFERENCES)
+	if (preferences.current_window != PREFERENCE_WINDOW_GAME_PREFERENCES)
 		return list()
 
 	var/static/list/admin_only_legacy_toggles = list(
