+++ code/modules/client/preferences_menu.dm
@@ -1,17 +1,20 @@
 /datum/verbs/menu/Preferences/verb/open_character_preferences()
 	set category = "OOC"
-	set name = "Open Character Preferences"
+	set name = "Characters"
 	set desc = "Open Character Preferences"
 
-	usr?.client?.prefs?.open_window(PREFERENCE_TAB_GAME_PREFERENCES_CHARACTER)
+	usr?.client?.prefs?.open_window(PREFERENCE_PAGE_CHARACTERS)
 
 /datum/verbs/menu/Preferences/verb/open_game_preferences()
 	set category = "OOC"
-	set name = "Open Game Preferences"
-	set desc = "Open Game Preferences"
+	set name = "Game Settings"
+	set desc = "Open Game Settings"
 
-	var/datum/preferences/preferences = usr?.client?.prefs
-	if (!preferences)
-		return
+	usr?.client?.prefs?.open_window(PREFERENCE_PAGE_SETTINGS)
 
-	usr?.client?.prefs?.open_window(PREFERENCE_TAB_GAME_PREFERENCES_SETTINGS)
+/datum/verbs/menu/Preferences/verb/open_volume_mixer()
+	set category = "OOC"
+	set name = "Volume Mixer"
+	set desc = "Open Volume Mixer"
+
+	usr?.client?.prefs?.open_window(PREFERENCE_PAGE_PREFERENCES_VOLUME)
